requirements.txt
huggingface-hub
langchain
langchain-huggingface
python-dotenv
speechrecognition
pyaudio
streamlit
pyttsx3
pycountry